Remove indices dir and move indices to tools
