+++ tests/indices/test_indices.py
@@ -4,7 +4,7 @@ import xarray as xr
 import numpy as np
 import pytest
 import os
-from climakitae.indices.indices import (
+from climakitae.tools.indices import (
     effective_temp,
     noaa_heat_index,
     fosberg_fire_index,
